+++ src/infra/jwt/jose/config.go
@@ -5,15 +5,19 @@ import (
 )
 
 type Config struct {
-	IssuerURL string
-	CacheTTL  time.Duration
-	Audience  []string
+	IssuerURL           string
+	CacheTTL            time.Duration
+	Audience            []string
+	CustomClaimPath     string
+	PermissionClaimPath string
 }
 
-func NewConfig(issuerURL string, audience []string, cacheTTL time.Duration) *Config {
+func NewConfig(issuerURL string, audience []string, customClaimPath, permissionClaimPath string, cacheTTL time.Duration) *Config {
 	return &Config{
-		IssuerURL: issuerURL,
-		CacheTTL:  cacheTTL,
-		Audience:  audience,
+		IssuerURL:           issuerURL,
+		CacheTTL:            cacheTTL,
+		Audience:            audience,
+		CustomClaimPath:     customClaimPath,
+		PermissionClaimPath: permissionClaimPath,
 	}
 }